+++ app/frontend/wallet/cardano-wallet.ts
@@ -8,7 +8,7 @@ import {MAX_INT32} from './constants'
 import shuffleArray from './helpers/shuffleArray'
 import NamedError from '../helpers/NamedError'
 import {Lovelace} from '../state'
-import {computeRequiredTxFee, selectMinimalTxPlan, isUtxoProfitable} from './byron-tx-planner'
+import {computeRequiredTxFee, selectMinimalTxPlan, isUtxoProfitable} from './byron/byron-tx-planner'
 import {MaxAmountCalculator} from './max-amount-calculator'
 // eslint-disable-next-line no-unused-vars
 import {TxPlan} from './shelley/build-transaction'
